round current price to 100s for easier price adj
Button by trade calc will now get the current price of bitcoin, rounded to the 100ths for easier organization. Without rounding, prices in the cents would be difficult to veiw when looking at the long list of trades, making it difficult to understand.

Button must be pressed to update. If it were live, it could change right before the start trade button is pressed, resulting in undesired trade prices being  set.