React.StrictMode Alright, look. I get the need to catch bugs that might not be noticed in development, but when the expected behavior is to run once, but the "helpful debug mode" runs it twice... Idk, that really really feels like a bug. Not a feature.
Like hear me out, what if something only works when it's run twice.
You may waste a lot of time thinking it's working fine, only to push
it to the test build stage and find out it doesn't work.

I understand that you should not buy a product as an effect, but not
everyone is building shopifoo 2.

Syntactically, "strict mode" should mean the behavior does strictly what
is expected of it. If you want to change the expected behavior of
something, call it "test mode".

Big thank you to myself for being the only person reading this <3